+++ CHANGELOG.md
@@ -4,12 +4,13 @@ Yii2 Gentelella Change Log
 1.2.0 under development
 -----------------------
 
+- Enh #7: Add a an ability to add custom items to dropdown list of `yiister\gentelella\widgets\Panel` (fps01)
 - New #8: Added a `yiister\gentelella\widgets\Timeline` widget (fps01)
 - New #5: Added an ability to keep a sidebar state (fps01)
 - Fix #6: using a PHP 5.4 instead of 5.5
 
 1.1.0 June 26, 2016
------------------------
+-------------------
 
 - New #3: Added a `yiister\gentelella\widgets\StatsTile` widget (fps01)
 - New #2: Added a `yiister\gentelella\widgets\grid\GridView` widget (fps01)